+++ src/main/java/pl/plajer/buildbattle/menus/options/registry/biomes/BiomesRegistry.java
@@ -18,6 +18,7 @@
 
 package pl.plajer.buildbattle.menus.options.registry.biomes;
 
+import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -75,11 +76,12 @@ public class BiomesRegistry {
   }
 
   private void registerInventory() {
-    Inventory inv = Bukkit.createInventory(null, Utils.serializeInt(biomes.size()),
+    Inventory inv = Bukkit.createInventory(null, Utils.serializeInt(biomes.size() + 1),
         plugin.getChatManager().colorMessage("Menus.Option-Menu.Items.Biome.Inventory-Name"));
     for (BiomeItem item : biomes) {
       inv.addItem(item.getItemStack());
     }
+    inv.addItem(Utils.getGoBackItem());
     inventory = inv;
   }